refactor project structure
