+++ src/manifest.json
@@ -10,8 +10,7 @@
   },
 
   "background": {
-    "scripts": [ "background.js" ],
-    "type": "module"
+    "scripts": [ "background.js" ]
   },
 
   "permissions": [
