+++ tsconfig.json
@@ -1,5 +1,5 @@
 {
-    "include": ["browser-extension/*"],
+    "include": ["src/*"],
     "compilerOptions": {
         "target": "es2022",
         "lib": ["es2022", "es6", "dom"],
